refactor datatypes adaptors
