--- aeon/datatypes/_adapter/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for pandas/dask conversion adapters."""
