+++ aeon/datatypes/_hierarchical/_check.py
@@ -86,7 +86,7 @@ check_dict[("pd_multiindex_hier", "Hierarchical")] = check_pdmultiindex_hierarch
 
 
 if _check_soft_dependencies("dask", severity="none"):
-    from aeon.datatypes._adapter.dask_to_pd import check_dask_frame
+    from aeon.utils.conversion.dask_to_pd import check_dask_frame
 
     def check_dask_hierarchical(obj, return_metadata=False, var_name="obj"):
         return check_dask_frame(
