+++ aeon/datatypes/_panel/_convert.py
@@ -1095,7 +1095,7 @@ _extend_conversions("numpy2D", "numpy3D", convert_dict, mtype_universe=TYPE_LIST
 
 
 if _check_soft_dependencies("dask", severity="none"):
-    from aeon.datatypes._adapter.dask_to_pd import (
+    from aeon.utils.conversion.dask_to_pd import (
         convert_dask_to_pandas,
         convert_pandas_to_dask,
     )
