+++ aeon/datatypes/_series/_convert.py
@@ -212,7 +212,7 @@ if _check_soft_dependencies("xarray", severity="none"):
 
 
 if _check_soft_dependencies("dask", severity="none"):
-    from aeon.datatypes._adapter.dask_to_pd import (
+    from aeon.utils.conversion.dask_to_pd import (
         convert_dask_to_pandas,
         convert_pandas_to_dask,
     )
