+++ aeon/utils/conversion/pd_multiindex_to_list_dataset.py
@@ -1,15 +1,8 @@
-from deprecated.sphinx import deprecated
+"""Convert from pr.DataFrame with MultiIndex to ListDataset."""
 
 from aeon.utils.validation._dependencies import _check_soft_dependencies
 
 
-# TODO: remove in v0.8.0
-@deprecated(
-    version="0.6.0",
-    reason="convert_from_multiindex_to_listdataset will be moved from datatypes in "
-    "v0.8.0 to utils.conversion in v0.8.0",
-    category=FutureWarning,
-)
 def convert_from_multiindex_to_listdataset(trainDF, class_val_list=None):
     """
     Output a dataset in ListDataset format compatible with gluonts.
